requirements.txt
transformers
rdkit
tqdm
matplotlib
scikit-learn
FPSim2
pyyaml
numpy
networkx
python-louvain
tape_proteins
huggingface_hub
meeko
prody
pandas